# pulse/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
host: localhost:6532
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/kevin/node-pulse
  description: |-
    Node Pulse is a distributed monitoring system for tracking node health, metrics, and alerts in real-time.

    ## Features
    - Real-time beacon heartbeat monitoring
    - Multi-metric data collection and analysis
    - Configurable alert rules with suppression
    - Webhook notifications for alerts
    - Historical data queries and export
    - Role-based access control (RBAC)

    ## Authentication
    Most endpoints require authentication. Use the `/api/v1/auth/login` endpoint to obtain a session token.
    Include the token in the `Authorization` header: `Bearer <your-token>`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Node Pulse API
  version: "1.0"
paths:
...
  /alerts/records:
    get:
      consumes:
      - application/json
      description: Retrieves alert records with optional filtering by node_id, level,
        status, and time range.
      parameters:
      - description: Filter by node UUID
        in: query
        name: node_id
        type: string
      - description: Filter by level (P0, P1, P2)
        enum:
        - P0
        - P1
        - P2
        in: query
        name: level
        type: string
      - description: Filter by status (pending, in_progress, resolved)
        enum:
        - pending
        - in_progress
        - resolved
        in: query
        name: status
        type: string
      - description: Filter from this time (ISO 8601)
        in: query
        name: start_time
        type: string
      - description: Filter until this time (ISO 8601)
        in: query
        name: end_time
        type: string
      - default: 50
        description: Maximum records to return (1-100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Records to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of alert records
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List alert records
      tags:
      - Alert Records
  /alerts/records/{id}/status:
    put:
      consumes:
      - application/json
      description: 'Updates the status of an alert record. Valid status transitions:
        pending→in_progress, in_progress→resolved.'
      parameters:
      - description: Alert record ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Alert record status updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or status transition
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Alert record not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update alert record status
      tags:
      - Alert Records
...
  /api/v1/admin/audit/logs:
    get:
      description: Query audit logs with filters and pagination
      parameters:
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        type: string
      - description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get audit logs (Admin)
      tags:
      - admin
  /api/v1/admin/audit/logs/:id:
    get:
      description: Get a specific audit log entry
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get audit log by ID (Admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"